api key security improvements 1
Moving values into parameter store to be loaded at runtime and remove them from the template/env vars.

The current API secret key is being deprecated and CommunityPatch will switch to RSA signing.
